refactor(dispatch): extract post-import policy cluster from core (D-review #4)
Operator-approved split (2026-07-17 post-review): the five post-import
search-policy appliers (_resolve_post_import_search_policy,
_apply_post_import_search_action, _apply_or_stage_transition/_denylists,
_run_or_stage_quality_gate) move verbatim to lib/dispatch/post_import.py
with their finalize_request DI seam. core.py returns under its
structural threshold (1078 -> 961 lines) and the decision-17/19 policy
surface now lives in one named module.
